% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twophase.R
\name{summaryTwoPhase}
\alias{summaryTwoPhase}
\title{summaryTwoPhase functions}
\usage{
summaryTwoPhase(eval_vec, se_vec, q, p, method)
}
\arguments{
\item{q}{threshold value for TPR(q)}

\item{p}{threshold value for FPR(p)}

\item{method}{"Benchmark", "ML"}

\item{Outcome}{vector of length n}

\item{X}{matrix of size n x t (t standard predictors included in the model)}

\item{Z}{matrix of size n x m (m biomarkers included in the model); some subjects have NA values}

\item{Stratum}{vector of length n}

\item{Phase_ID}{vector of length n}

\item{namesX}{names of X; vector of length t}

\item{namesZ}{names of Z; vector of length m}

\item{numBoot}{number of bootstrap samples}
}
\value{
return the summary. TPR,FPR,AUC together with their standard error estimates
}
\description{
summaryTwoPhase functions
}
\examples{
summaryTwoPhase(est_ml, se_ml, q=0.2, p=0.2, method="ML")
}
